+++ SDC/Model/InvestoreOwnerShapeModel.swift
@@ -12,7 +12,7 @@ class InvestoreOwnerShape{
     
 
     var clientNo, securityID, securityIsin, Security_Reuter_Code,Security_Name , Pending_In, Pending_Out, Security_Close_Price, Quantity_Owned ,Quantity_Avilable, Quantity_Pledge , Quantity_Freezed, Quantity_Unlisted , Security_Sector_Desc, Nominal_Value , Security_Sector  : String?
-    var  MarketValue : Int?
+    var  MarketValue : Double?
 
     init(data: [String: Any]) {
         
@@ -80,7 +80,7 @@ class InvestoreOwnerShape{
             self.Security_Name = Security_Name
         }
         
-        if let MarketValue = data["MarketValue"] as? Int {
+        if let MarketValue = data["MarketValue"] as? Double {
             self.MarketValue = MarketValue
         }
